requirements.txt
pyautogui==0.9.54
keyboard==0.13.5
pillow>=10.0.0
